requirements.txt
Flask==2.2.2
# Flask-SQLAlchemy==3.0.3
# Flask-Migrate==4.0.4
Pillow==9.4.0